New attachment manager and a couple of updates.
I created a manager portlet for attachments. I decided to not go
the route of using the existing portlet because I didn't want to
break the current usage. But that doesn't mean I'm opposed to such
a wrote in the future. The portlet lists all the available
attachments from the database. This can be used to find things that
you have already attached as there is currently no mechanism for
that. I also included a mechanism for previewing pdfs, though it
doesn't work on firefox for some reason. Problem with the upstream
library.

I also updated the version of fine-uploader. I didn't bother yet
to integrate it with the existing system, I just wanted to get this
off the ground first.

Changed the name of the file in logback to SimpleContentPortlet to
match the repo. Makes it easier for some of my utility scripts.