Fineuploader cleaning.
There was an extra step on the fineuploader of deleting the
responseJSON.success for no appearent reason. I tried removing it
and everything seemed fine.

Also removed, PDFObject's files since it is being brought in via
maven. I also updated the original code base to use the latest
version of FineUploader and it worked, so I removed those files
from the code base as well.