Pagination support enabled.
Turns out the portal doesn't like displaying all 1600+ attachments
at once. This is largely due to the way it fetches all of them.
Specifically it fetches all of them with their data even if I'm
not going to be using that data. So it tried to load like 2gb of
stuff to display the manager. What I've done now is enabled
pagination so that it only shows 100 at a time. It seems to be
working now.